chore: move all search to hook
